+++ packages/backend/src/schemas/v2.ts
@@ -1,6 +1,6 @@
 // THIS IS A AUTO-GENERATED FILE. DO NOT MODIFY MANUALLY!
 
-export type Repos = GitHubConfig | GitLabConfig | GiteaConfig;
+export type Repos = GitHubConfig | GitLabConfig | GiteaConfig | LocalConfig;
 
 /**
  * A Sourcebot configuration file outlines which repositories Sourcebot should sync and index.
@@ -151,5 +151,25 @@ export interface GiteaConfig {
      * List of individual repositories to exclude from syncing. Expected to be formatted as '{orgName}/{repoName}' or '{userName}/{repoName}'.
      */
     repos?: string[];
+  };
+}
+export interface LocalConfig {
+  /**
+   * Local Configuration
+   */
+  type: "local";
+  /**
+   * Path to the local directory to sync with. Relative paths are relative to the configuration file's directory.
+   */
+  path: string;
+  /**
+   * Enables a file watcher that will automatically re-sync when changes are made within `path` (recursively). Defaults to true.
+   */
+  watch?: boolean;
+  exclude?: {
+    /**
+     * List of paths relative to the provided `path` to exclude from the index. .git, .hg, and .svn are always exluded.
+     */
+    paths?: string[];
   };
 }