+++ packages/backend/src/utils.ts
@@ -1,5 +1,6 @@
 import { Logger } from "winston";
 import { AppContext, Repository } from "./types.js";
+import path from 'path';
 
 export const measure = async <T>(cb : () => Promise<T>) => {
     const start = Date.now();
@@ -15,9 +16,9 @@ export const marshalBool = (value?: boolean) => {
     return !!value ? '1' : '0';
 }
 
-export const excludeForkedRepos = (repos: Repository[], logger?: Logger) => {
+export const excludeForkedRepos = <T extends Repository>(repos: T[], logger?: Logger) => {
     return repos.filter((repo) => {
-        if (repo.isFork) {
+        if (!!repo.isFork) {
             logger?.debug(`Excluding repo ${repo.id}. Reason: exclude.forks is true`);
             return false;
         }
@@ -25,9 +26,9 @@ export const excludeForkedRepos = (repos: Repository[], logger?: Logger) => {
     });
 }
 
-export const excludeArchivedRepos = (repos: Repository[], logger?: Logger) => {
+export const excludeArchivedRepos = <T extends Repository>(repos: T[], logger?: Logger) => {
     return repos.filter((repo) => {
-        if (repo.isArchived) {
+        if (!!repo.isArchived) {
             logger?.debug(`Excluding repo ${repo.id}. Reason: exclude.archived is true`);
             return false;
         }
@@ -35,7 +36,7 @@ export const excludeArchivedRepos = (repos: Repository[], logger?: Logger) => {
     });
 }
 
-export const excludeReposByName = (repos: Repository[], excludedRepoNames: string[], logger?: Logger) => {
+export const excludeReposByName = <T extends Repository>(repos: T[], excludedRepoNames: string[], logger?: Logger) => {
     const excludedRepos = new Set(excludedRepoNames);
     return repos.filter((repo) => {
         if (excludedRepos.has(repo.name)) {
@@ -59,4 +60,17 @@ export const getTokenFromConfig = (token: string | { env: string }, ctx: AppCont
 
 export const isRemotePath = (path: string) => {
     return path.startsWith('https://') || path.startsWith('http://');
+}
+
+export const resolvePathRelativeToConfig = (localPath: string, configPath: string) => {
+    let absolutePath = localPath;
+    if (!path.isAbsolute(absolutePath)) {
+        if (absolutePath.startsWith('~')) {
+            absolutePath = path.join(process.env.HOME ?? '', absolutePath.slice(1));
+        }
+
+        absolutePath = path.resolve(path.dirname(configPath), absolutePath);
+    }
+
+    return absolutePath;
 }